+++ karios/report/product_generator.py
@@ -61,7 +61,10 @@ def _to_feature(series: Series, geo_transform: tuple, properties: list[str]) -> 
         "geometry": {"type": "Point", "coordinates": [x, y]},
         # 1* allow convert np.float32 to float that
         # allow json serialization (serialize np.float32 fail)
-        "properties": {prop: 1 * series[prop] for prop in properties},
+        # nan set to none for good json serialisation
+        "properties": {
+            prop: None if np.isnan(series[prop]) else 1 * series[prop] for prop in properties
+        },
     }
 
 
@@ -85,17 +88,24 @@ class ProductGenerator:
         - mask if `gen_kp_mask` (-kpm)
         - KP Raster if `gen_delta_raster (-gip)`
         - KP geojson if reference image have projection
+
+        Returns:
+            list(str): list of generated product paths
         """
+        product_paths = []
         if self._config.gen_kp_mask:
-            self._create_mask()
+            product_paths.append(str(self._create_mask()))
 
         if self._config.gen_delta_raster:
-            self._create_intermediate_raster()
+            product_paths.append(str(self._create_intermediate_raster()))
 
+        # always generate JSON if inputs products is geo referenced
         if not self._reference_image.get_epsg():
             logger.warning("Unable to generate KP GeoJSON, reference image not geo referenced")
         else:
-            self._create_kp_geojson()
+            product_paths.append(str(self._create_kp_geojson()))
+
+        return product_paths
 
     def _create_intermediate_raster(self):
         logger.info("Create KP raster product")
@@ -117,13 +127,17 @@ class ProductGenerator:
         dx_band_array[y_index, x_index] = self._points["dx"]
         dy_band_array[y_index, x_index] = self._points["dy"]
 
+        output_file_path = os.path.join(self._config.output_directory, "kp_delta.tif")
+
         self._reference_image.to_raster(
-            os.path.join(self._config.output_directory, "kp_delta.tif"),
+            output_file_path,
             [dx_band_array, dy_band_array],
             gdal.GDT_Float32,
         )
 
         logger.info("KP raster product created")
+
+        return output_file_path
 
     def _create_mask(self):
         logger.info("Create KP product mask")
@@ -134,19 +148,25 @@ class ProductGenerator:
             [self._reference_image.y_size, self._reference_image.x_size], dtype=np.uint8
         )
         final_mask[y_index, x_index] = 1
-        self._reference_image.to_raster(
-            os.path.join(self._config.output_directory, "kp_mask.tif"), final_mask
-        )
+        output_file_path = os.path.join(self._config.output_directory, "kp_mask.tif")
+        self._reference_image.to_raster(output_file_path, final_mask)
         logger.info("KP product mask created")
+        return output_file_path
 
     def _create_kp_geojson(self):
         logger.info("Create KP vector product")
+
+        # configure properties to export in features
+        columns_to_export = ["dx", "dy", "score", "radial error", "angle"]
+        if "zncc_score" in self._points.columns:
+            columns_to_export.append("zncc_score")
+
         # creates feature for each dataframe rows
         feature_as_series = self._points.apply(
             _to_feature,
             axis=1,
             geo_transform=self._reference_image.geo_transform,
-            properties=["dx", "dy", "score", "radial error", "angle"],
+            properties=columns_to_export,
         )
 
         feature_collection = {
@@ -163,3 +183,5 @@ class ProductGenerator:
             out.write(json.dumps(feature_collection, indent=3))
 
         logger.info("KP vector product created")
+
+        return output_file